apps/api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
redis==5.0.1
httpx==0.27.0
supabase==2.3.0
anthropic>=0.40.0
